backend/requirements.txt
httpx
pydantic
pytest
python-dotenv
alpaca-trade-api
respx
pytest-asyncio
pydantic-settings
black
flake8
yfinance
pandas-ta
fastapi
uvicorn